reorganise tests to modularise functionality and dependencies, close #249, restructure CI
