+++ .github/workflows/python-tests-non-pipeline.yml
@@ -1,4 +1,4 @@
-name: Tests
+name: Python tests, non-pipeline
 
 on: [workflow_call]
 
@@ -29,5 +29,5 @@ jobs:
           key: ${{ runner.os }}-poetry-tests-${{ hashFiles('poetry.lock') }}
       - name: Install project
         run: poetry install --no-interaction --sync -E test
-      - name: Run unit tests on Python ${{ matrix.python-version }}
-        run: poetry run pytest -vv
+      - name: Run NON-pipeline unit tests on Python ${{ matrix.python-version }}
+        run: poetry run pytest -vv --ignore=tests/pipeline
