+++ .github/workflows/python-tests-pipeline.yml
@@ -1,0 +1,33 @@
+name: Python tests, pipeline
+
+on: [workflow_call]
+
+jobs:
+  tests:
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.10"]
+        os: [ubuntu-20.04, ubuntu-22.04]
+    runs-on: ${{ matrix.os }}
+
+    steps:
+      - uses: actions/checkout@v3
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Set up Poetry
+        uses: snok/install-poetry@v1    # v1.3.4 at time of adoption
+        with:
+          poetry-version: "1.7.1"
+      - uses: actions/cache@v3
+        with:
+          path: |
+            ~/.cache/pypoetry/virtualenvs
+            ~/.cache/pytest
+          key: ${{ runner.os }}-poetry-tests-${{ hashFiles('poetry.lock') }}
+      - name: Install project
+        run: poetry install --no-interaction --sync -E test -E pipeline
+      - name: Run PIPELINE unit tests on Python ${{ matrix.python-version }}
+        run: poetry run pytest -vv tests/pipeline
